changes required statement for information when used
the sentence says that the callbacks are not required when participant cannot
finish immediatelly. This is part that I do understand. It's up to me as a participant
implementator to ensure that participant is or isn't capable to finish right when asked by coordinator.
But the second part talks about coordinator is unable. That's out of control
me as an implementator of the participant. That's why I think removing
'required' is appropriate.